Cloud PINAC-Nexus First Preview
This is the first preview of pinac-nexus deployed on Vercel. Removed the python server for the code, removed login system and also removed input box for api keys in profile Page. Currently we have only Gemini in cloud, so we have removed ChatGPT for model options. Currently this cloud server lack chat history, we will fix this in upcoming updates.