Adds Python backend
(For now, this is just a sample backed for testing)

Introduces a Flask-based Python backend. This includes setting up API endpoints for status checks and data processing. The backend can be run in both development and production environments, utilizing waitress for production performance. It also introduces backend.spec, and requirements.txt to handle venv, build, dependencies, and
PyInstaller configurations. The Electron app is updated to spawn and manage the Python backend process, enabling communication between the frontend and backend.