Uses Python env  for build process
Ensures the build process uses the Python executable from the virtual environment.

This change prevents issues where the system's Python interpreter is used instead of the venv's, leading to incorrect dependencies or version conflicts. It also adds error handling to check for the existence of the venv and provides informative messages to the user if it's missing.